Functions that are also in lib-cove - delete them.
We don't want to duplicate them and they aren't used here anyway.

(With one exception - get_file_type is needed here, so a requirement for lib-cove is added)

That isn't to say that lib-cove must be the final place for this code; we may realise other considerations means that code should move around later.
But the code not being duplicated will make sure that it only lives in one place and any moves later are done carefully.

We also remove UnrecognisedFileTypeXML even tho that's not in lib-cove yet; but it depends on something that does, and it should probably be in IATI lib anyway.